Day 17 pt 1 solution
this seems like it _should_ solve part 2, as well, and it would, if I had the hours and hours to run it. but otherwise, not so much. there's 1 swing at optimization, by keeping my "used spaces" set small, but even so, that's not going to cut it, apparently. Since I'm officially on vacation, maybe I'll look tomorrow!