this is part 1 done, and part 2 attempted but not yet complete
this got a lot easier when I stopped trying to parse it with 1 big regex, and intead used a couple smaller ones so i didn't have to know how many #'s i was reading in.

part two I'm currently "stuck" on getting cards off the end of the card-list, but i'm tired, i'll figure it out tomorrow over lunch or after dinner.