day 5 pt 1
here's part 1's solution. it has a theoeretical solution for pt2 ,except hte #'s are too big for the computer to tolerate, so we're going to have to actually deal w/ ranges, rather than individual values. It's gonna have to wait till i have more time to think about it.