Day 6
fun and exciting map navigation! I expanded my Direction enum by giving it the ability to turn left and right, and made a Coord struct with 2 ints, because tuples aren't hashable. i also made a new Motion struct which is a position and a facing direction which helped a lot tracking the guards motion, and probably will be useful later, if what happened in prior years holds true.

the "trick" for part 2 is that if the new obstacle isnt' in the guards original path, it will obviously never get in the guards way, and thus be a failure, so that greatly reduces our number of possibilities. Still a bit slow to check em all, though, ~2 secs.