Day 7
aaah, recursion!! But also, whew, recursion. my poor little brain could not see that as the solution at first. I might not have needed to generate the possibilities first, and that wouldve been faster, but this way i got to re-use the orders for additional lists of the same size, so... win some, lose some?